android-tools: skip recipes with lower version
Issue: LINCD-18964

There are two sets of android-tools related recipes in meta-oe, the
lower version set is located in recipes-devtools directory, and the
higher version set is located in dynamic-layers/selinux/recipes-devtools
directory. This will cause the meta-selinux layer check to fail in
yocto-check-layer.

The root cause is the lower version of android-tools is used when there
is no meta-selinux layer, while the higher version of android-tools is
used when the meta-selinux layer is added, which results in different
signatures. Since these recipes are not whitelisted in wrlinux, skip the
lower version via SKIP_RECIPE to avoid the layer checking failure is a
simple approach.

(LOCAL REV: NOT UPSTREAM) -- WRLinux specific

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>