check-version-mismatch.bbclass: handle non-elf executables
Issue: LINCD-23515

Using qemu to run non-elf executables such as shell scripts directly
is destined to fail. In such case, we check its interperter and try
out best to run it accordingly.

We'll also need to skip the "/etc" directory as files under it are
configuration files and init scripts. And the init script will
send SIGTERM and SIGKILL to all processes, giving users annoying
behavior.

(LOCAL REV: NOT UPSTREAM) -- sent to meta-oe on 20250417

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>