conf/version-check.conf: skip piglit
Issue: LINCD-23515

piglit is an open-source test suite for OpenGL implementations. It
contains a whole lot of programs which all do not give out any version
information. Running version check for piglit is a waste of time (~1h),
so setting a '%' to skip it.

(LOCAL REV: NOT UPSTREAM) -- sent to meta-oe on 20250417

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>