python3-aiohttp: fix CVE-2025-53643
Issue: LIN1025-2676

AIOHTTP is an asynchronous HTTP client/server framework for asyncio and
Python. Prior to version 3.12.14, the Python parser is vulnerable to a
request smuggling vulnerability due to not parsing trailer sections of
an HTTP request. If a pure Python version of aiohttp is installed (i.e.
without the usual C extensions) or AIOHTTP_NO_EXTENSIONS is enabled,
then an attacker may be able to execute a request smuggling attack to
bypass certain firewalls or proxy protections. Version 3.12.14 contains
a patch for this issue.

References:
https://nvd.nist.gov/vuln/detail/CVE-2025-53643

(LOCAL REV: NOT UPSTREAM) --  send to meta-openembedded on 20250716

Signed-off-by: Jiaying Song <jiaying.song.cn@windriver.com>