hdf5: fix CVE-2025-2913
Issue: LIN1025-199

A vulnerability was found in HDF5 up to 1.14.6. It has been
rated as problematic. Affected by this issue is the function
H5FL__blk_gc_list of the file src/H5FL.c. The manipulation of
the argument H5FL_blk_head_t leads to use after free. An
attack has to be approached locally. The exploit has been
disclosed to the public and may be used.

References:
https://nvd.nist.gov/vuln/detail/CVE-2025-2913
https://git.openembedded.org/meta-openembedded/commit/?h=scarthgap&id=b42e6eb3e51ae73aed0012f1ca827ff30e34f436

Upstream patch:
https://github.com/HDFGroup/hdf5/commit/7cc8b5e1010a09c892bc97ac32d9515c3777ce07

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>