hdf5: fix CVE-2025-2914
Issue: LIN1025-198

A vulnerability classified as problematic has been found in HDF5
up to 1.14.6. This affects the function H5FS__sinfo_Srialize_Sct_cb
of the file src/H5FScache.c. The manipulation of the argument sect
leads to heap-based buffer overflow. Local access is required to
approach this attack. The exploit has been disclosed to the public
and may be used.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2025-2914
https://git.openembedded.org/meta-openembedded/commit/?h=scarthgap&id=7d1b63f0af74a354f08de2b683e98ccd7f357b05

Upstream patch:
https://github.com/HDFGroup/hdf5/commit/804f3bace997e416917b235dbd3beac3652a8a05

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>