hdf5: fix CVE-2025-2915
Issue: LIN1025-197

A vulnerability classified as problematic was found in HDF5 up
to 1.14.6. This vulnerability affects the function H5F__accum_free
of the file src/H5Faccum.c. The manipulation of the argument
overlap_size leads to heap-based buffer overflow. Attacking locally
is a requirement. The exploit has been disclosed to the public
and may be used.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2025-2915
https://git.openembedded.org/meta-openembedded/commit/?h=scarthgap&id=bd847d489a4ca3538b2315b061f0f591df0e8ec9

Upstream patch:
https://github.com/HDFGroup/hdf5/commit/26a76bafdef3a0950d348a08667de161a19b7c2c

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>